fix(static): fix spread detection to use latest value of attributes, ensure native color parses sub-styles
