+++ packages/web/types/helpers/normalizeColor.d.ts
@@ -1,4 +1,9 @@
 import * as NCC from '@tamagui/normalize-css-color';
 export declare const rgba: typeof NCC.rgba;
+export declare const names: {
+    transparent: number;
+    white: number;
+    black: number;
+};
 export declare const normalizeColor: (color?: string | null, opacity?: number) => string | undefined;
 //# sourceMappingURL=normalizeColor.d.ts.map
